backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
openai
python-dotenv